% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_chromatogram.R
\name{annotate_chromatogram}
\alias{annotate_chromatogram}
\alias{list_IDs}
\title{Annotate chromatograms}
\usage{
annotate_chromatogram(chr_data, sample_info)

list_IDs(chr_data)
}
\arguments{
\item{chr_data}{Chromatogram data imported via \code{read_unicorn}}

\item{sample_info}{Sample information data frame.}
}
\description{
Merge additional information about the sample conditions into a chromatogram.
}
\details{
UNICORN sometimes codes sample information into file names using a funny 
naming scheme. In such cases, adding sample information during import via 
\code{read_unicorn} is not straightforward. An easier approach is to parse 
the data without sample names, list sample IDs, build a data frame 
accordingly, and finally merge it with the chromatogram.

The information is supplied in a data frame called \code{sample_info}. This
data frame must have at least two columns: \code{ID} and \code{Sample} (case 
sensitive). Merging is done based on the \code{ID} column. \code{Sample} and 
all other columns are added accordingly. Missing \code{ID}s in 
\code{sample_info} get an \code{NA} as a sample name but this is reported as
as warning.
}
\note{
This function may overwrite some of the annotation from chromatogram
data if \code{sample_names} was already used during import via \code{read_unicorn}.
all other columns are be added accordingly.
}
\seealso{
\code{\link{read_unicorn}}
}

