% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{convert_values}
\alias{convert_values}
\title{Convert values}
\usage{
convert_values(x, y, z)
}
\arguments{
\item{x}{Date for the initial entry. \code{x} must be in ymd format, e.g., "2017-03-30".}

\item{y}{Value for the initial entry.}

\item{z}{Date for the target entry. \code{z} must be in ymd format, e.g., "1988-09-23".}
}
\value{
Return the inflation incorporated price of the initial price \code{y}, given the initial date \code{x} and target date \code{z}.
}
\description{
Convert values
}
\examples{
> convert_values("2017-03-30", 100, "1988-09-23")
[1] 49.04768
}
