% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Find peaks}
\usage{
find_peaks(chr_curve, window_factor = 0.001, threshold_factor = 0.01,
  verbose = FALSE)
}
\arguments{
\item{chr_curve}{A subset of chromatogram data frame that corresponds to a
singe curve}

\item{window_factor}{A coefficient to be used in determining the peak
search window size. Default is \code{0.001}.}

\item{threshold_factor}{A coefficient to be used in filtering out baseline-
level peaks (i.e. noise). Default is \code{0.01}.}

\item{verbose}{Prints progress during the run. Default is \code{FALSE}.}
}
\description{
Find local maxima in a chromatogram
}
\details{
This function searches for local maxima along the curve of a chromatogram by
dividing it into smaller windows. It works on a single curve at a time,
therefore the data must be grouped before calling this function.

Window size is defined by \code{window_factor}, which is a multiple of
the total run volume. It determines the sensitivity of the search: A smaller
window size may result in too many false peaks but it runs faster. A larger
window size is more robust but it slows down the calculation. The default
value (\code{0.001}) corresponds to 1/1000th of the total run volume and it
is a fairly good starting point to search for peaks.
}
\examples{
\dontrun{
peaks <- find_peaks(chr_data, window_factor = 0.001, threshold_factor = 0.01)
}

}

