% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{inflation.through.time}
\alias{inflation.through.time}
\title{Plot Consumer Price Index Through Time}
\usage{
inflation.through.time(x, y)
}
\arguments{
\item{x}{Date for the initial entry}

\item{y}{Date for the final entry}
}
\value{
Plot the calculated CPI for All Urban Consumers \code{x} until \code{y}.
}
\description{
Plot Consumer Price Index Through Time
}
\examples{
> inflation.through.time("1975-01-01","1985-01-01")
}
