% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing.R
\name{unicorn}
\alias{unicorn}
\title{Show CSV Head or Null Values and Preprocess Data}
\usage{
unicorn()
}
\description{
This function allows users to input the path of a CSV file and then choose to view the head of the data,
count the null values in each column, or preprocess the data.
}
