% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_chromatogram.R
\name{annotate_chromatogram}
\alias{annotate_chromatogram}
\title{Annotate chromatograms}
\usage{
annotate_chromatogram(chr_data, sample_info)
}
\arguments{
\item{chr_data}{Chromatogram data imported via \code{read_unicorn}}

\item{sample_info}{Sample information data frame.}
}
\description{
Merge additional information about the sample conditions into a chromatogram
}
\details{
UNICORN sometimes codes sample information in file names using a very 
non-intuitive naming scheme. In such cases, adding sample information during
import via \code{read_unicorn} is not straightforward. An easier approach is
then to parse the data without sample names, list sample IDs, build a data 
frame accordingly, and merge it with the chromatogram.

The information is supplied in a data frame called \code{sample_info}. This
data frame must have at least two columns: \code{ID} and \code{Sample} (case 
sensitive). Merging is done based on the \code{ID} column. \code{Sample} and 
all other columns are be added accordingly.
}
\note{
This function may overwrite some of the annotation from chromatogram
data if \code{sample_names} was used during import via \code{read_unicorn}.
}

