% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Find peaks}
\usage{
find_peaks(chr_data, maximum_peaks = 1)
}
\arguments{
\item{chr_data}{Chromatogram data imported by \code{read_unicorn} or 
\code{read_res}}

\item{maximum_peaks}{Maximum number of peaks to find.}
}
\description{
Find local maxima in a chromatogram
}
\details{
This function searches for local maxima along the curves of chromatograms. 
Because the search can yield even slightest spikes as peaks, the hits are 
sorted in a descending order based on the UV absorbance the the top 
\code{maximum_peaks} (default 1) are reported.

The algorithm is modified from the \code{findPeaks} function in 
\href{https://CRAN.R-project.org/package=quantmod}{quantmod} package.
}
\examples{
\dontrun{
peaks <- find_peaks(chr_data, 5)
}

}
