% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_unicorn.R
\name{read_unicorn}
\alias{read_unicorn}
\title{Import chromatograms}
\usage{
read_unicorn(file_name, sample_names = NULL, combined = TRUE,
  single_channel = FALSE, hardware_autozero = FALSE,
  reference_measurement = 1, verbose = FALSE)
}
\arguments{
\item{file_name}{The file to be imported. It can be either an Excel file or
an ASC (currently not implemented).}

\item{sample_names}{A vector or a data frame containing actual sample names
corresponding to curve names}

\item{combined}{\code{TRUE} (default) when several spectra are combined
by the UNICORN software prior to export.}

\item{single_channel}{\code{TRUE} if the UV cell can only work in single
channel mode. Default is \code{FALSE}.}

\item{hardware_autozero}{\code{TRUE} when the UV cell was autozeroed during
acquisition. Default is \code{FALSE}.}

\item{reference_measurement}{The number of measurement to be used in
baseline subtraction. Ignored when \code{hardware_autozero = TRUE}.}

\item{verbose}{Display additional progress information. Default is
\code{FALSE}.}
}
\value{
A tibble data frame that contains volume, (baselined) absorbance,
wavelength/channel, curve and sample name information for the experiment.
}
\description{
Read exported UNICORN chromatogram data
}
\details{
If \code{sample_names} is a character vector, a new column named as
\code{Sample} will be appended to the data frame \emph{in the same order} as
curve names. However, this is not recommended as the order of the curves may
change or some curves may be excluded at some point during data import. In
the latter case, you will get an error message. The safest way is to define a
data frame with the column names \code{ID} and \code{Sample}, matching the
curve name to the sample name, respectively. When you use it to define sample
names, only those curves that are found in the data will be processed, i.e.
missing curves will be ignored.

When \code{hardware_autozero = TRUE}, chromatograms are assumed to be
baselined already during acquisition and this baseline level will is trusted.
Otherwise (default), the baseline will be subtracted from each chromatogram
using the measurement defined by \code{reference_measurement}. This is a
linear operation, which will not change the shape of the curve but it will
only shift its baseline towards 0.

Please note that \code{single_channel} refers to the UV cell rather than how
the data are exported. Even if you export only one channel out of many from a
chromatogram file, the naming will include the channel information and it
will interfere with data import as well as grouping. Therefore, use
\code{single_channel = TRUE} only when you are sure that the UV detector can
operate with one channel at a time.
}
\examples{
\dontrun{
samples <- data.frame(
    ID = c("Manual run 10:1", "Manual run 10:5", "Manual run 10:3"),
    Sample = c("Negative control", "Positive control", "Test"))
chr_data <- read_unicorn("Chromatogram.xls", samples)
}

}
\seealso{
\code{\link{annotate_chromatogram}} for annotating chromatograms
without sample information, \code{\link{list_IDs}} to see parsed curve 
names, \code{\link{create_annotation}} to create a data frame with sample
information.
}

